Added support files back into their module files because import was being annoying
