+++ KTANExpert.py
@@ -13,9 +13,6 @@ from modules.morseModule import *
 from modules.compWires import *
 from modules.wireSeq import *
 
-from modules.support.morse import *
-from modules.support.who2lists import *
-
 
 #Def Vars
 isDoing = True
