--- modules/support/who2lists.py
@@ -1,30 +1,0 @@
-#Copy this to make it easier: ", ". So basically WORD CtrlV WORD CtrlV WORD and so on and so forth
-
-ready = ["yes", "okay", "what", "middle", "left", "press", "right", "blank", "ready"]
-first = ["left", "okay", "yes", "middle", "no", "right", "nothing", "uhhh", "wait", "ready", "blank", "what", "press", "first"]
-no = ["blank", "uhhh", "wait", "first", "what", "ready", "right", "yes", "nothing", "left", "press", "okay", "no"]
-blank = ["wait", "right", "okay", "middle", "blank"]
-nothing = ["uhhh", "right", "okay", "middle", "yes", "blank", "no", "press", "left", "what", "wait", "first", "nothing"]
-yes = ["okay", "right", "uhhh", "middle", "first", "what", "press", "ready", "nothing", "yes"]
-what = ["uhhh", "what"]
-uhhh = ["ready", "nothing", "left", "what", "okay", "yes", "right", "no", "press", "blank", "uhhh"]
-left = ["right", "left"]
-right = ["yes", "nothing", "ready", "press", "no", "wait", "what", "right"]
-middle = ["blank", "ready", "okay", "what", "nothing", "press", "no", "wait", "left", "middle"]
-okay = ["middle", "no", "first", "yes", "uhhh", "nothing", "wait", "okay"]
-wait = ["uhhh", "no", "blank", "okay", "yes", "left", "first", "press", "what", "wait"]
-press = ["right", "middle", "yes", "ready", "press"]
-you = ["sure", "you are", "your", "you're", "next", "uh huh", "ur", "hold", "what?", "you"]
-youare = ["your", "next", "like", "uh huh", "what?", "done", "uh uh", "hold", "you", "u", "you're", "sure", "ur", "you are"]
-your = ["uh uh", "you are", "uh huh", "your"]
-youre = ["you", "you're"]
-ur = ["done", "u", "ur"]
-u = ["uh huh", "sure", "next", "what?", "you're", "ur", "uh uh", "done", "u"]
-uhhuh = "uh huh"
-uhuh = ["ur", "u", "you are", "you're", "next", "uh uh"]
-whatq = ["you", "hold", "you're", "your", "u", "done", "uh uh", "like", "you are", "uh huh", "ur", "next", "what?"]
-done = ["sure", "uh huh", "next", "what?", "your", "ur", "you're", "hold", "like", "you", "u", "you are", "uh uh", "done"]
-next = ["what?", "uh huh", "uh uh", "your", "hold", "sure", "next"]
-hold = ["you are", "u", "done", "uh uh", "you", "ur", "sure", "what?", "you're", "next", "hold"]
-sure = ["you are", "done", "like", "you're", "you", "hold", "uh huh", "ur", "sure"]
-like = ["you're", "next", "u", "ur", "hold", "done", "uh uh", "what?", "uh huh", "you", "like"]
